refactor: WorldTourScreen to use TopBar and update strings
