+++ presentation/ui/src/main/res/values-ar/strings.xml
@@ -2,7 +2,7 @@
 <resources>
     <string name="world_tour">جولة حول العالم</string>
     <string name="country_name">أسم البلد</string>
-    <string name="start_exploring_the_world_movie_by_enter_your_favorite_country_in_search_bar">ابدأ استكشاف أفلام العالم عن طريق إدخال دولتك المفضلة في شريط البحث</string>
+    <string name="country_tour_description">ابدأ استكشاف أفلام العالم عن طريق إدخال دولتك المفضلة في شريط البحث</string>
     <string name="country_tour">جولة حول البلاد</string>
     <string name="no_search_result">لا توجد نتائج للبحث</string>
     <string name="please_try_with_another_keyword">يرجى المحاولة بكلمة مفتاحية أخرى</string>
