+++ presentation/ui/src/main/res/values/strings.xml
@@ -2,7 +2,7 @@
 <resources>
     <string name="world_tour">World Tour</string>
     <string name="country_name">Country name</string>
-    <string name="start_exploring_the_world_movie_by_enter_your_favorite_country_in_search_bar">Start exploring the world movie by enter your favorite country in search bar</string>
+    <string name="country_tour_description">Start exploring the world movie by enter your favorite country in search bar</string>
     <string name="country_tour">Country tour</string>
     <string name="no_search_result">No search result</string>
     <string name="please_try_with_another_keyword">please try with another keyword.</string>
